feat: enhance blog post page layout and functionality
Refactor the BlogPostPage component to improve the layout by
introducing a new main structure, including a header and footer.
Update the BlogPostContent component to display blog details. 
Replace the existing HTML content parsing logic and streamline
data retrieval with improved parameter handling.
